Add Parser and FileToFileParser
Adds a base Parser class, from which, other parser types can be derived.
Also adds the FileToFileParser, which reads from a file and outputs to
another file, which then replaces the original if changed.

This speeds up processing as lines to be processed are not cached in
memory before being written in one chunk.